+++ src/cli/ArgumentsBuilder.php
@@ -29,6 +29,7 @@ final readonly class ArgumentsBuilder
                 $argv,
                 'hv',
                 [
+                    'diff',
                     'help',
                     'version',
                 ],
@@ -44,6 +45,7 @@ final readonly class ArgumentsBuilder
         }
 
         $file    = null;
+        $diff    = false;
         $help    = false;
         $version = false;
 
@@ -51,6 +53,11 @@ final readonly class ArgumentsBuilder
             assert(is_array($option));
 
             switch ($option[0]) {
+                case '--diff':
+                    $diff = true;
+
+                    break;
+
                 case 'h':
                 case '--help':
                     $help = true;
@@ -74,6 +81,7 @@ final readonly class ArgumentsBuilder
 
         return new Arguments(
             $file,
+            $diff,
             $help,
             $version,
         );